github integration, projet terminé

+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/RuleNameRepository.java
@@ -4,5 +4,5 @@ import com.nnk.springboot.domain.RuleName;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RuleNameRepository extends JpaRepository<RuleName, Integer> {
+public interface RuleNameRepository extends JpaRepository<RuleName, Long> {
 }
